backend/requirements.txt
fastapi
numpy
pandas
opencv-python
pillow
python-multipart
uvicorn